+++ docs/about/whats_new.rst
@@ -8,6 +8,7 @@ What's New
 Next Release
 ============
 
+- Add `dataset count` cli command and deprecate `datacube-search` :pull:`1963`
 - docs: remove broken link :pull:`1973`
 - examples: remove unimplemented zeros :pull:`1974`
 - examples: fix function signatures :pull:`1972`